Rename to nlpo3, make it shorter, less prone to typo [cd build]
