+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "pythainlp-rust-modules"
+name = "nlpo3"
 version = "1.1.0"
 edition = "2018"
 license = "Apache"
@@ -7,13 +7,10 @@ authors = ["Thanathip Suntorntip Gorlph"]
 include = ["pyproject.toml"]
 
 [lib]
-name = "oxidized_thainlp_main_lib"
 path = "src/lib.rs"
 # "cdylib" is necessary to produce a shared library for Python to import from.
-#
 # Downstream Rust code (including code in `bin/`, `examples/`, and `tests/`) will not be able
-# to `use string_sum;` unless the "rlib" or "lib" crate type is also included, e.g.:
-# crate-type = ["cdylib", "rlib"]
+# to `use string_sum;` unless the "rlib" or "lib" crate type is also included.
 crate-type = ["cdylib", "rlib"]
 
 [workspace]
@@ -24,6 +21,8 @@ ahash = "0.7.2"
 binary-heap-plus = "0.4.1"
 bytecount = "0.6.2"
 lazy_static = "1.4.0"
+rayon = "1.5"
+regex = "1.4.6"
 smol_str = "0.1.17"
