+++ README.md
@@ -1,6 +1,7 @@
-# oxidized-thainlp
+# nlpo3 (formerly oxidized-thainlp)
 
-Thai Natural Language Processing in Rust, with Python-binding.
+Thai Natural Language Processing in Rust,
+with Python and Node bindings.
 
 ## Features
 
@@ -20,7 +21,7 @@ pip install pythainlp-rust-modules
 
 Use:
 ```python
-from oxidized_thainlp import load_dict, segment
+from nlpo3 import load_dict, segment
 
 load_dict("path/to/dict.file", "dict_name")
 segment("สวัสดีครับ", "dict_name")
