+++ nodejs_binding/.gitignore
@@ -1,2 +1,2 @@
 node_modules/
-oxidized_nodejs_thainlp/*.node
+nlpo3/*.node
