+++ nodejs_binding/README.md
@@ -1,11 +1,6 @@
-# oxidized-nodejs-thainlp
+# nlpo3
 
 Like the root directory, but with nodejs binding.
-
-## Features
-
-- newmm dictionary-based word tokenization, at ultra fast speed
-- support custom dictionary
 
 ## Build It Yourself
 
@@ -19,46 +14,43 @@ Like the root directory, but with nodejs binding.
 ```bash
 # In this directory
 npm run release
+```
 
+Before build, your `nlpo3` directory should look like this
 ```
-Before build, your oxidized_nodejs_thainlp should look like this
-```
-- oxidized_nodejs_thainlp
+- nlpo3
     - rust_mod.d.ts
     - index.ts
-
 ```
+
 After
 ```
-- oxidized_nodejs_thainlp
+- nlpo3
     - rust_mod.d.ts
     - rust_mode.node
     - index.ts
     - index.js
-
 ```
+
 ## Usage
 
-For now, copy the whole oxidized_nodejs_thainlp directory after build to your project 
+For now, copy the whole `nlpo3` directory after build to your project 
 
 In NodeJS:
 ```javascript
-const  oxidizedThaiNLP = require(`${path_to_oxidized_nodejs_thainlp_directory}`)
+const oxidizedNLP = require(`${path_to_nlpo3}`)
 
-oxidizedThaiNLP.loadDict("path/to/dict.file", "dict_name")
-oxidizedThaiNLP.segment("สวัสดีครับ", "dict_name")
+oxidizedNLP.loadDict("path/to/dict.file", "dict_name")
+oxidizedNLP.segment("สวัสดีครับ", "dict_name")
 ```
 
 In TypeScript (run on node)
-
 ```typescript
-
-import {segment,loadDict} from `${path_to_oxidized_nodejs_thainlp_directory}/index`
+import {segment,loadDict} from `${path_to_nlpo3}/index`
 
 loadDict("path/to/dict.file", "dict_name")
 segment("สวัสดีครับ", "dict_name")
 ```
-
 
 Just that!
 
@@ -70,6 +62,3 @@ Please report issues at https://github.com/PyThaiNLP/oxidized-thainlp
 # TODO
 
 Find a way to build and publish on npm.
-
-
-
