+++ nodejs_binding/nlpo3/rust_mod.d.ts
@@ -1,3 +1,2 @@
-
 export function segment(text: string, dict_name: string, safe: boolean, parallel: boolean): string[];
 export function loadDict(file_path: string, dict_name: string): string;
