+++ python_binding/Cargo.toml
@@ -1,22 +1,16 @@
 [package]
-name = "oxidized-thainlp"
+name = "nlpo3-python"
 version = "1.1.0"
 edition = "2018"
 license = "Apache"
 authors = ["Thanathip Suntorntip Gorlph"]
 
 [lib]
-name = "oxidized_thainlp"
 path = "src/lib.rs"
-# "cdylib" is necessary to produce a shared library for Python to import from.
-#
-# Downstream Rust code (including code in `bin/`, `examples/`, and `tests/`) will not be able
-# to `use string_sum;` unless the "rlib" or "lib" crate type is also included, e.g.:
-# crate-type = ["cdylib", "rlib"]
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-pythainlp-rust-modules = { path = ".." }
+nlpo3 = { path = ".." }
 
 [dependencies.pyo3]
 version = "0.13.2"
